LayoutTests/imported/w3c/web-platform-tests/navigation-api/per-entry-events/dispose-cross-document-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: Invalid key
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: Invalid key


Harness Error (TIMEOUT), message = null

TIMEOUT No dispose events are fired due to cross-document forward pruning Test timed out
